requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
python-dotenv==1.0.0
whitenoise==6.6.0
requests
django-ipware


